Create README.md rwadzano
I would like to build an upclass ecd grade 1 to 7 center for the underprivileged children in the community.  I also want to include the parents in our day to day work so they can be able to continue helping children at home. I would also like to start a feeding scheme for them as they come from impoverished families.